refactor(platform): move payment, reporting, and audit records out of everruns-core
Carve the backend accounting/observability records out of the runtime-facing
`everruns-core` contract and into `everruns-platform`, preserving every
persisted and API wire format (types move verbatim with identical serde/utoipa
derives). Follows the EVE-837 pattern: platform -> core only, never the reverse.

Moved to everruns-platform:
- payment records: PaymentAccount, PaymentPolicy, PaymentAttempt, plus the
  PaymentOwnerType / PaymentStatus value enums they embed.
- reporting: the whole module (ReportQuery/ReportResult/DatasetCatalog/… and the
  ReportingProjectionSink / ReportingQueryBackend traits).
- audit: the whole module (AuditEvent/AuditAction/…, the AuditLogger and
  HasAuditTargetId traits, and their fixtures).

Stays in everruns-core (structurally anchored, re-exported from platform):
- Payment execution contract PaymentRail, PaymentMethod, MachinePaymentRequest,
  MachinePaymentResponse — bound to core's `PaymentAuthority` trait and
  `ToolContext`; moving them would force a forbidden core -> platform edge.

Consumers (server routes/services/stores/openapi registration) now import the
moved types from `everruns_platform`. Worker is untouched (only uses the
execution contract, which stays in core). No new inter-crate everruns edge is
introduced, so publish/pin wiring is unchanged.

Fixes EVE-838
