+++ crates/server/src/domains/agent_triggers/commands.rs
@@ -29,13 +29,14 @@ use crate::storage::models::{
 use chrono::Utc;
 use everruns_core::typed_id::{AgentId, SessionId, TriggerId};
 use everruns_core::{
-    AgentAction, AgentIdentityId, AgentTrigger, AgentTriggerType, AuditEvent, Caller,
-    InvocationSessionMode, Policy, ScheduleTriggerConfig,
+    AgentIdentityId, AgentTrigger, AgentTriggerType, Caller, InvocationSessionMode, Policy,
+    ScheduleTriggerConfig,
 };
 use everruns_durable::{
     CreateScheduleRow, Pagination as DurablePagination, ScheduleExecutionFilter,
     ScheduleTargetType, StoreError, UpdateField, UpdateSchedule, WorkflowEventStore,
 };
+use everruns_platform::{AgentAction, AuditEvent};
 use serde::Deserialize;
 use serde_json::{Value, json};
 use std::str::FromStr;
