+++ crates/server/src/openapi.rs
@@ -637,27 +637,27 @@ fn schema_extensions_mut(schema: &mut Schema) -> Option<&mut Option<Extensions>>
             api::knowledge_indexes::ListKnowledgeIndexesQuery,
             api::knowledge_indexes::KnowledgeIndexResponse,
             api::knowledge_indexes::KnowledgeIndexDocumentResponse,
-            everruns_core::payment::PaymentAccount,
-            everruns_core::payment::PaymentAttempt,
-            everruns_core::payment::PaymentPolicy,
+            everruns_platform::payment::PaymentAccount,
+            everruns_platform::payment::PaymentAttempt,
+            everruns_platform::payment::PaymentPolicy,
             everruns_core::payment::PaymentRail,
-            everruns_core::payment::PaymentStatus,
-            everruns_core::payment::PaymentOwnerType,
+            everruns_platform::payment::PaymentStatus,
+            everruns_platform::payment::PaymentOwnerType,
             domains::payments::types::CreatePaymentAccountRequest,
             domains::payments::types::UpdatePaymentAccountRequest,
             domains::payments::types::CreatePaymentPolicyRequest,
             domains::payments::types::UpdatePaymentPolicyRequest,
-            everruns_core::reporting::DatasetCatalog,
-            everruns_core::reporting::DatasetCatalogEntry,
-            everruns_core::reporting::ReportColumn,
-            everruns_core::reporting::ReportColumnKind,
-            everruns_core::reporting::ReportFilter,
-            everruns_core::reporting::ReportFilterOp,
-            everruns_core::reporting::ReportOrderBy,
-            everruns_core::reporting::ReportOrderDirection,
-            everruns_core::reporting::ReportQuery,
-            everruns_core::reporting::ReportResult,
-            everruns_core::reporting::ReportTimeRange,
+            everruns_platform::reporting::DatasetCatalog,
+            everruns_platform::reporting::DatasetCatalogEntry,
+            everruns_platform::reporting::ReportColumn,
+            everruns_platform::reporting::ReportColumnKind,
+            everruns_platform::reporting::ReportFilter,
+            everruns_platform::reporting::ReportFilterOp,
+            everruns_platform::reporting::ReportOrderBy,
+            everruns_platform::reporting::ReportOrderDirection,
+            everruns_platform::reporting::ReportQuery,
+            everruns_platform::reporting::ReportResult,
+            everruns_platform::reporting::ReportTimeRange,
             ListResponse<domains::reporting::types::SavedReport>,
             domains::reporting::types::CreateSavedReportRequest,
             domains::reporting::types::DatasetProjectorLag,
